requirements.txt
anyio==3.6.2
Authlib==1.2.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
exceptiongroup==1.1.1
fastapi==0.95.2
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.1
pluggy==1.0.0
prometheus-client==0.17.0
prometheus-fastapi-instrumentator==6.0.0
psycopg2==2.9.6
pycparser==2.21
pydantic==1.10.8
pytest==7.3.1
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
timeout-decorator==0.5.0
tomli==2.0.1
typing_extensions==4.6.2
urllib3==2.0.2
uvicorn==0.22.0
